requirements.txt
PyGithub==1.53
seaborn==0.11.0
wordcloud==1.8.0
